Add more test messages.
Crucially, I also changed the type of a column in
the Users table.  I think that perhaps 45 characters
was not enough for the hashes, which was causing
them to not match.  I changed the type to VARCHAR(100).
So far still broken, though.  I'm not sure why the hashes
aren't matching up.