Remove LSNA broken section "Services Rendered".
The Services Rendered section and report is broken.  It was
not intended to be placed on the production site, since it was
not yet finished, but had been committed to CH's repository
at the time when that repository was used to create a new
production site.  I deleted this section from the participant profile,
deleted the link from the reports landing page, and deleted the report
file in this commit.