Rename contactus.php to contact.php, and spam-protect help address.
The rename is largely a matter of taste, but I think it's generally
good to keep URLs short, and keep the number of words in a page name
to a minimum (to reduce cognitive load).  But the real reason is just
that I always call this page "contact" everywhere, and I'll keep
typing it wrong if it's something different here :-).

The alternative spam-protected email address seems wise, based on past
experience (but see issue #9).