Script now bidirectional.
This version determines whether to hash or restore a plaintext password
based on whether the retrieved information matches a stored hashed password.
Unfortunately, I see that this assumes that (1) we will have the current stored
password and (2) it will already be hashed.  While (1) is doable, (2) sort of
defeats the purpose.  We could do the check the other way, such that we check the
plaintext (definitely plaintext) passwords against the stored passwords, and then
it's the stored passwords that are or are not hashes.  In that case we need to read
in 2 lists of passwords.  Thoughts?