Continue testing download script.
Not sure what happened last time.  I got the "The File does not exist"
message from my testing print lines, but the script also resulted in a downloaded
file.  It was called download.php, though, not test.txt as it should have been.

I will run is_file in the short test version on the same file path as the longer,
actually-meant-to-download version and see what happens.