Continue testing download script.
The last commit had success in downloading the contents of
the pointed-to file.  In this commit, I attempt to point to a
file with the desired information, rather than a test file.

I had to manually add the path in download.php, which is something
we'll want to change - either all downloads should be in the same
global directory (probably) or we need another get to choose a
directory by group name.