Fix table and column names in update_passwords.php.
Since the start of the SQL output says 'use ttm-core;', we don't need
to repeat the table name in the individual UPDATE statements.

And the column we want to use to select the row to update is the
username.  Due to issue #14, that column is currently called
"User_Email", but it's really a username.  When issue #14 is fixed,
we'll just update this script.