Put PHPass 0.3 distribution into include/phpass-0.3/.
See also commit 3940a04883294c3d3349bfb5e16655ea40a4a28e, which this
partially re-enacts (that commit was on the 'phpass' branch, which was
merged to and later unmerged from master).  It's okay to put this on
master now because it doesn't change any TTM code; once we start
modifying code, that will happen on its own branch, which will draw
selectively from the previous 'phpass' branch work.