For regular login, allow hashed or plain passwords in DB.
This is for issue #2, and is the first step in the gradual
reintroduction of certain changes from the `phpass` branch.

Note that the same shim introduced here -- the one that allows either
hashed or non-hashed password in the database -- needs to be done in
ajax/change_password.php too, as that file is still doing a direct
string comparison to verify the old password before allowing the new
password to be set.  Thanks to @cecilia-donnelly for noticing this.