Fix dosage percentage bug.
This commit is not ready to be merged.  It creates a
new file, dosage_percentage.php, which calculates
dosage percentage for display on both the program
profile and inclusion in exports.  It has not yet
been tested on dev.  At the moment, the results
should display correctly on the profile but not
in production_exports.php yet.  The export requires
more information to be displayed, and it isn't
clear how/whether to calculate the sum of hours in
dosage_percentage.php.