Test dosage percentage fix.
Modify dosage_percentage calculation for use in
production_exports.php.  If successful, this commit
will show correct dosage percentage on program
profile and in program exports.
Before pulling this commit to the dev server, I
will create a test case with the same characteristics
as the bug visible on the production server.