Fix hours display.
The last commit fixed the displayed dosage
percentage on the program profile but
had the unintended consequence of making all
"Total hours in this program" and "Total
hours across funded programs" values equal
to zero.
In this commit I used the second return value
from the dosage_calculation function for the
"total hours in this program" column.  I
kept the "Total hours across funded programs"
column the same, except to use the same
returned value for the hours in the currently
selected program.