Fix export of dosage percentage and hours.
Still not ready to merge this branch. This
commit moves all the elements from the
calculate_dosage return array into the fputcsv
array in production_exports.php.  After this
commit, the return elements should appear correctly
in the export file.
This will allow us to see whether the hours are
calculating correctly in the function, so we can more
accurately diagnose the problem in the display
on the program profile (see the previous commit
message).