Testing messages: "max hours" or begin/end times?
I added some messages to see whether the problem stems from programs having "max
hours" equal to zero (or null, or some such) or whether the calculations are off
because of the beginning and end times recorded for the programs.  I added an
initial check for programs that have no hours at all entered.  For those programs,
the student will be said to have zero hours.