Change begin/end to default to db value.
Rather than testing both am and pm, I changed the code to test only for pm and
if not pm, then to default to the saved value instead of defaulting to zero.
Initially I tested for am, then pm (in which case I added 12) and then if neither
am nor pm matched, set the value to zero.  This seemed to be having the effect
of all beginning and ending times being read as zero.