Fix issue #32: restore user access privileges.
This fix works by restoring the code's semantics to what they were
before commit 4650c8730f3e504605a677db9e7a8e57f3aa9afe, but it is
almost certainly not The Right Fix.

The claim in commit 4650c8730f3e504605a677db9e7a8e57f3aa9afe that the
value of the cookie (implying the value of the "user" field) shouldn't
matter is clearly wrong -- it must matter, since this commit fixes the
bug.  But the fact that wrapping the username value in single quotes
makes a difference feels very, very yucky.  Why would that work but
the plain username not work?  Is that value from the cookie directly
interpolated into an SQL query somewhere, and the code is unknowingly
depending on those single quotes for SQL string syntax, for example?

The separate issue #33 is about getting to the bottom of this matter.