Add testing output to fix column order in Enlace production_exports.php
After the last commit, the Baseline_Assessments questions are now in the correct question/value pairs
in the .csv.  However, the order of the export columns is not consistent with the query I see.
This commit adds testing output to see how the headers of the export are being created.