Fix column order and remove testing output.
As of this commit the columns should now be in the correct order for the "New Surveys
Export" identified version.  The column order when ordered by In_Table was different
depending on whether we selected Baseline_Assessment_Question_ID or just the
Question column, which led to the disconnect between headers (in the Question
column order) and the values (following the Baseline_Assessment_Question_ID
order).