Pass connection parameter to SQL sanitization calls in trp/.
In commit 118eb7a4a7e2c and commit d949b15e3f7 I'd forgotten that
mysqli_real_escape_string() takes a connection variable; this commit
fixes that.  This is part of the work on issue #19.

For the most part, the fix is straightforward: add the $cnnTRP
parameter to the calls, and where necessary include dbconnopen.php
earlier so that $cnnTRP is available.  In a few places, things got
more complex, because the call to open the database wasn't nearby, or
was entangled in way, or whatever.  Below are places where something
non-straightforward happened in addition to the expected changes:

* trp/programs/profile.php, trp/ajax/save_gold_scores.php,
  trp/ajax/delete_participant.php: Just include one dbconnopen.php at
  the start of the file and one dbconnclose.php at the end, and get
  rid of all the ones in between.

* trp/participants/view_parent_survey.php: Remove what appears to be a
  redundant second inclusion of dbconnopen.php.

* trp/participants/profile.php, trp/ajax/search_users.php: Use cnnLISC
  for SQL-sanitizing the query that goes to the cnnLISC connection.

* trp/include/import_grades.php: Keep an existing connection open by
  simply removing an interior call that closed it.