Remove testing output and make each legend cell contain all possible values.
The way I did it in the last commit, each possible response showed up in its
own cell.  This was not helpful because each question has just one cell per column,
and the legend should appear in the cell directly beneath the question, with all
the possible responses in that cell.  As of this commit it should do that
as desired.