Add download script file for use in all exports.
I created a download script as a function in
generalized_download_script.php.  I am not sure whether
the headers will work as expected if used inside a function, since
headers require nothing to have been printed before they are called.
This file also includes a test call to the function, so we can see
whether it works before adding the complications of calling it from
another location.  This is the first step to finishing Issue #1.