Add testing output to see why mysql table contents did not appear.
When navigated to directly, the script worked correctly for a non-logged-in user
as of last commit (i.e. returned the error page and didn't download anything).
It worked fairly well for a logged-in user, except that it didn't retrieve the
information correctly.  It did, however, download a csv file with the correct
column headers!
As of this commit, I am adding some testing output.  I suspect that either the
dbconnopen file is not being included correctly or the query is not being
read correctly.
Issue #1.