Add another testing POST.
It's not clear to me why a set of post variables would be a problem, so I added
an extra test POST which is the exact copy of the de-identified post.  That
de-identified post is already reaching the response, so the copy of it with
 the empty set of variables should too.  If not, then perhaps the set of
variables to be posted must have at least one element?  The "identified" version
will have all the post variables sent.  With luck it too will reach a response
and I will get a better understanding of what is going wrong.

Issue #1.