Test both download types.
Removing the test output didn't solve the problem.  It may be that the headers
will not work with a POST from the export.php file.  For this one, I am
testing the download with the POST variable and the hardcoded variable
on a call from the export.php file.
Issue #1.