Add access check for download.
This commit adds a check for the logged-in users privilege list.  It checks
the Users_Privileges table to see whether the user is permitted to download
the specified file (i.e., someone who only had access to bickerdike would
not be permitted to download a file that was linked to the lsna database).

This is the next step in solving Issue #1.