Remove hardcoded DB names.
For Issue #46, remove hardcoded DB names in a few files.  In the files changed
in this commit, the hardcoding introduced a bug and the processes should
work correctly using just the table name.  Hardcoded schemas remain in two
other files that have a more complicated problem - see comments on Issue #46.