Use strtolower to set directory name to lowercase.
Issue #1.  I kept the test output here, but I suspect the problem in the last
few commits was that the connection file was not being found correctly,
because the DB name "LSNA" was in caps while the directory name "lsna" is
lowercase.
This might be something to look into standardizing.