Display La Casa as program on participant profile.
This commit moves the preliminary La Casa section to be included as a
program on the participant profile.  It also updates the SQL file
creating La Casa as a program to also connect it to our sample participants.
If the participant and program are not linked in the Participants_Programs
table then the program will not show up on that participant's profile.
Issue #5.