Create program profile for La Casa.
This commit creates a program profile section for La Casa.  To do this,
it also dries some of the code by pulling out the Program Enrollment section
which is repeated across programs.  This is a very preliminary commit - I
will see how it looks then refine.
Note: this commit also hardcodes the La Casa program ID as 6 in the
sql/change_trp_program.sql file.
Issue #5.