Display college credits over time - La Casa.
Show college change over time as requested by a user in a conversation
on August 22.  This commit introduces the display change on the participant
profile and starts the create statement for a new table in sql/la_casa_
create_tables.sql.  It isn't complete, but I'm committing it midway so I
can move on to another issue.
Issue #5.