Remove testing output and test.
As of commit b22d4f66, logging in worked correctly without single quotes.
Access privileges worked as covered in Issue #37.  A new participant
with an apostrophe in her name could be added as in Issue #60.  This
commit removes the testing output in ajax/login_submit.php to ensure that
all these fixes are still valid.

Issue #33.