Continue Issue #5 changes for TRP.
Switched a few more of the exports over to the download script for TRP.
NOTE: I know there is an error in the file  as it stands -- the TRP
downloads should have 'TRP' as the value for the 'db' array key, but
at the moment have 'SWOP'.  I'm committing here so that I can delete the
NMMA exports in their own commit (see Issue #51 for why I am deleting these).
They are already deleted in the Issue-51 branch, but I want to clean them
up here in their own commit.