Continue completing simple Issue #1 fixes on Enlace and TRP.
As of this commit, most exports that don't require special treatment have been converted
to the new, secure download script.  A few still remain on Enlace.

Of those that do require special treatment:

Note that TRP needs work on the Gold Scores export.  The existing export  will not work
smoothly for the new script, so we need to do some thinking about how to bring that
over.  Might be worth talking to Vito.

I added the legends argument to the script.  It assumes that the legends will come in as an
array, which does not quite work for the Enlace legends.  I could write a script to print
those as arrays for now.  After this commit I am going to try to merge the Issue #40 branch,
which added those legends, into a new Issue-1 branch to attempt to reconcile them.

Or, add a 'legends_query' argument that tells the script what query to execute to get
the correct Enlace legends.

Enlace's program-specific access levels are also an issue, mainly for the participants export.
How should I deal with that with the new script?  I could give the script two possible queries
to run, depending on access.