Continue Issue #1 fixes (finish on TRP).
As of this commit, all groups except Enlace have been switched over to
the new system for downloads/exports.  TRP still needs to be tested.
I hit a bit of a snag with Enlace, through my own fault.  I was
changing the wrong file -- all_exports.php instead of
production_exports.php.  It shouldn't be an enormous setback, since
most of the code is the same, but I will need to go back through it
and make sure that I'm using the correct queries.  After this I may
delete all_exports.php -- it's just causing trouble now.  I created it
to do some block group work in Fall 2013, and that is no longer
needed.  I still need to determine how I will take care of the legends
for Enlace's survey exports, but that shouldn't take long.  Almost done!