Finish conversion to downloads - some problems remain.
 All user-facing files now link to the download script for exports.  I
 included a non_admin_string element in the array to hold pieces of
 queries that are only relevant for non-admin users,  though testing
 is needed.

Next steps:
1. Test all downloads for TRP and Enlace, as well as representative
exports from other groups to make sure that changes have not broken
them.
2. Add  dosage calculation for the export in the download file.  We
need the results from the query before we can calculate it.

3. Replace the legends for the Enlace survey exports.  The original fix
will not work in the new system.  I will probably include the query as
an array element and calculate the legends in the script.