Fix array of downloads.
As of the last commit ( ), none of the downloads on the site were
working, even those that had been working previously.  I looked back
through commits and found that as of commit 65f871a the exports
stopped working.  I suspected that the problem was in the array of
download names, not in the script, since it appeared that the file
wasn't even finding the function (when not logged in, that page should
show an error message and didn't).  One inner array was missing a
close paren, which this commit adds.  With luck, this will resolve the
download problem.