Issue #1: Fix Enlace typos and uncomment.
I added some mising commas and single quotes in the Enlace array
values, so with luck this commit will see all downloads working as
expected.  If not, there are more errors in the Enlace download
names.  I also removed the testing output, since I am now sure that
the bug is caused by typos in the download_name array.