Fix enlace export bugs -- titles, legends, dosage.
Issue #1.  This commit adds testing output to debug  the titles and
legends for Enlace intake, impact, and combined ("new") surveys.  It
also adds testing output to resolve the dosage calculation issue (the
dosage results aren't appearing in the export).

It removes the query testing output and fixes the institution and
program attendance queries.