Test calculating dosage without pushing it to the array.
Issue #1.  After the last commit (a047a90), the browser couldn't find
the dosage exports.  This suggests that the if statement is now
working correctly, but something within it is causing a problem in the
creation of the file.  This commit comments out a few lines within the
if statement and adds testing output.