Issue #5: Reduce inputs for La Casa section of TRP.
A user and I discussed keeping the La Casa section very lean for now,
and in this commit I removed a lot of information from the participant
profile in favor of only showing the credits accrued and loans taken
on by the person shown in the profile.  I also altered the table
Create statements, since we will be tracking much less information
than originally anticipated.

I expect users will want us to add much of that information back in,
but the create statements and html tables for the profile will always
be there in history.

In order for this commit to work properly, I will need to load the new
tables in the db.