Issue #1. Fix Enlace bugs in exports.
Add non-admin string to impact surveys and alter the access string in
the intake and "new" surveys to test on whether the person is in the
allowed program, not whether the allowed program entered the survey.
That is, if Person X is in Program Y, then Program Y's staff should be
able to see all of Person X's surveys, even if they were entered by
Program Z.  To that end, I changed the query to test on the link
between the person and the allowed program (Program Y) rather than on
the program that entered the survey (Program Z).