Issue #1. Fix Enlace surveys so that participant surveys are visible to users from all programs they are involved in.
I thought I fixed this in 0a73c939, but it turns out that I should
have used Session_Names.Program_ID instead of
Participants_Programs.Program_ID in order to get the desired results.
This is a problem related to the Session ID v. Program ID
confusion... the so-called "Program_ID" in Participants_Programs is
actually a Session ID, but the access is referring to real programs,
not sessions, so I had to look for the program id in Session_Names,
where the actual program id is stored as Program_ID.