Issue #1. Fix Enlace surveys - add correct program links for new surveys.
The new surveys did need that first non-admin string, in order to link
people to the correct programs so that the program-specific users
will be able to export their surveys.  That is, the query had been
linking programs based solely on the program the survey was entered
in, and it should have been linking people to programs based on the
Participants_Programs table.