Issue #5.  Draft of participant entry from program profile complete.
As of this commit, users can add participants from the La Casa program
page and edit the information pertaining to those participant on their
profile pages.  In the next draft I'll need to add a dropdown for race
and likely a date popout.  I also plan to add household income,
household size, and race/ethnicity reports to the program page.