Issue #78.  Add site-specific authz check.
This commit adds an TRP-specific authorization check and saves it
in trp/siteconfig.php.  This file will be included in all TRP
files and will check authorization before allowing a file to load.
This commit also adds a function to /ajax/session_test.php that checks
the access level, and removes this site-specific check from auth.php.