Issue #78.  Add permission level check to session_test.php
This commit adds a new function to test the level of access for a
logged-in user and return that access level.  If the logged-in user is
not an administrator, then siteconfig.php uses the value of the
function return to set either a $VIEW_ONLY or $DATA_ENTRY variable to
indicate which elements should be hidden for that user.  Following
commits will use conditional statements to show or hide elements based
on the logged-in user's permission/access level.