Issue #78.  Show/hide TRP elements based on access level.
This commit adds conditional checks around TRP elements that were
designed not to be shown to users with data entry and/or view-only
access.  These elements had previously been hidden using the
"hide_on_view" and "no_view" classes.  As of this commit, they are
placed inside "if" statements that check whether the logged-in user is
a view only user ($VIEW_ONLY is set to 1 in siteconfig.php) or a data
entry user ($DATA_ENTRY is set to 1 in siteconfig.php).