Issue #78. Rewrite TRP access check to be more readable.
This commit rewrites the admin/data entry/view only access check to
use one single variable ($AccessLevelTRP) for all access levels, and
compares it to variables named with words (e.g. $AdminAccess) instead
of magic numbers 1, 2, and 3.

It covers all areas that had been hidden with HTML classes in previous
code iterations, but in testing I realized that there are some parts
of the code that should be access-limited and are not (see e.g. the
"Add New Session" section in trp/programs/profile.php?id=1).