Issue #78. Add program-specific access check.
This commit adds a function to ajax/session_test.php that checks the
logged-in user's access to specific programs.  This is only relevant
for TRP and Enlace (so far).  I'll have to consider how to allow a
single user access to multiple programs, but it shouldn't be difficult
on this end.  The function returns an array.  Right now, it assumes
that it will find one or many rows in the Users_Privileges table.