Issue #5: Add fields to data entry on participant profile.
This commit adds remaining fields to the participant profile for both
editing and display.  It also changes the types of some columns in the
La_Casa_Basics table and adds queries to the
ajax/save_la_casa_info.php file to correspond to the new fields on the
participant profile.