Issue #5: Add first insertion of constant data.
As of the last commit, the so-called "constant" data in the first La
Casa table on the participant profile could not be inserted into the
La_Casa_Basics table, but only updated.  This commit fixes that by
checking to see whether the participant has a row in the
La_Casa_Basics table, then creating one if not.  This also makes the
change such that every row about a participant in the LC Basics table
will have the same "constant" information.  We are assuming these
items will not change, so do not track their change over time.