Issue #80: Change new survey export.
This commit changes the new survey export/report to connect to the
session and program names via a session id stored in the Assessments
table.  It also adds a SQL file which adds that column to the
Assessments table and provides for it to be filled with previous
information from the Participants_Caring_Adults table, if we choose to
solve the backwards-compatibility issue that way. (see the issue for
more discussion)