Issue #77: Write function to find dosage sum.
This commit adds a function that finds the sum of hours for a given
participant, for both programs and mentorship time.  It needs more
testing.  The next commit will call it from the "export" page for all
participants. It might be included in the existing dosage export, as
another column.