Issue #80: Fix queries.
Fix trivial error in /enlace/classes/participant.php query.  See line
note
https://github.com/OpenTechStrategies/lisc-ttm/commit/b909db3c293b74e2cd41bf4425d6b7b9ab7fc487#commitcomment-9724021
on @xray7224's commit b909db.

@xray7224 and I co-wrote a query in enlace/programs/profile.php that
turns out to be incorrect.  We grouped on Participants.Participant_ID,
but that has the unintended effect of showing each participant only
once per program, even though each participant may correctly appear in
multiple sessions.  I've corrected the query in this commit by
grouping first on the session and then on the participant id.

I also updated the join to the Assessments table so that it joins on
both the participant id and the session id, as @xray7224 and I
discussed in IRC on 12 Feb 2015, around 10:32am UTC-6.  This prevents
non-related assessments from being counted, so that our "intake survey
completed" column is accurate.