Issue #119: Add date range to involvement report.
This commit removes the bar chart from the LSNA extent of involvement
report and adds a date range filter.  The bar chart isn't appearing
correctly on production, and notes in issue #119 suggest that adding
more features to the table of results will be more useful to users
than the bar chart, in any case.