Issue #119: Add participant names in dropdown spans.
As requested in the issue, this commit adds an option to view the
names of the people who are referred to by the counts in this report.
To do that, I wrote a function that creates a span with a list of
those names, and sets the span to toggle visibility on click of the
number displayed.

Since the two previous commits, 1b087d4 and 812c4a1, added posting to
the same page I changed the report menu links so that those pages show
up as themselves on refresh.  Since I set the pages to show up as an
entire page, not as a div in the middle of the page, I added headers
and footer to participant_roles.php.

It also cleans up some of the opening php markers in the LSNA header
file and removes commented out information and obsoleted information
in the reports.php file.

A lot of the changes in program_involvement.php are actually
indentation fixes.  Sorry for the ugly diff!  I did remove the results
by program for now, with the intention of adding them back in a
cleaner way in a commit to come.