Issue #78: Add function to check access to multiple programs.
Add function enforce_access_program_array(), which checks whether the
logged-in user has access to one of a set of programs.  This is useful
on several Enlace pages, where a user has permission to see the page
based on whether s/he has access to a program.  For example, the user
may have access to a given participant profile if the participant is
enrolled in a program that the user has access to.

This commit also adds the new access checks to the programs
subdirectory.