Issue #78: Resolve user class conflict in Bickerdike subsite.
Instead of a participant class, like the other subsites, Bickerdike
has a user class.  This obviously causes a conflict with our new
access user class (and global $USER), so I renamed it to
Bickerdike_User wherever a new instance is invoked. I also removed a
subdirectory named "deletes" that appeared to be full of deprecated
files.