#78: Convert LSNA dirs to new access API.
This commit adds the new access API to the classes, institutions, and
participants subdirectories of the LSNA subsite.  It also adds it to
the LSNA header and index files.  I removed the old cookie-dependent
CSS styling in the header and in the files where it appeared (notably
the "no_view" class on many elements.

Some of the pages had a deprecated setup where instead of loading a
new page for the profile, search, and add new pages they showed and
hid divs.  Those divs had long since been deprecated in favor of
loading each individual page, but the headers of the pages still had
the old show/hide javascript.  I removed that in this commit.

Similarly, some links to other pages were needlessly using the
"set_participant_id.php" and similar pages.  Previously, these links
had set a cookie to show a certain div, but now they are obsolete.  I
removed them from some of these pages as I was cleaning up.

I also changed the logout target here to point to the central logout
page, which will correctly end the session.  This should be done
across subsites.