#78: Use new access check on LSNA program directory.
This commit checks access at the top of program pages and also nests
elements in "if" statements that should only be shown to qualified
users.  This still needs a serious test with a read-only user.

I also removed two deprecated files in this commit.  I removed
category_profile.php on master, too, in commit 25ebde3.