#78: Use new access check on LSNA reports directory.
This commit adds the new access check to the head of files in the
LSNA reports directory.  It uses include_once() instead of include
because many of the files are included within lsna/reports/reports.php
and shown or hidden via divs.

This commit also removes some deprecated lines including javascript
and jqplot.  Those are not needed since those pages don't have charts
created with jqplot on them.

Finally, it fixes a couple errors introduced in the previous LSNA
commits (81125ad, 38c9392).  I changed the header links to plain
links, instead of the previous system of setting cookies to show
various pages and fixed one "if" statement that was missing the "if."