#117: Add funder to each program and campaign event.
This commit also adds a new "funders.php" report which filters events
by date and funder(s).

Note that the changes to program_profile.php happen twice because the
schedule is duplicated.  It appears once for programs and once for
campaigns.  Deduplicating it seemed like a more complicated fix, so I
just made the change in both places for now.