#5: Update tables and add new fields.
The LC table structure needed to updated to account for the fact that some
(many) fields will change every school term.  I added the "LC_Terms" table to
account for that, and changed a good deal of the add/show/report on participant
stuff accordingly.  This still needs to be fully tested before I send it out to
users, unfortunately, but it's in good shape now.