#5: Corrections to display and edit.
Not all constant data was correctly saved and editable, but now those queries
have been fixed.  This commit also improves the creation of dropdowns for
constant data to include the current value, so data isn't accidentally lost on
edit.  It also fixes some display issues with the report on the program profile.