#78: Alter links to SWOP participant profiles.
This is a stopgap measure, and I'd welcome advice on a better way to fix this.
The problem: participants can be represented by either
participants/pool_profile.php or participants/participant_profile.php, depending
on whether or not they're in the pool.  Often we want to link to a participant
without necessarily knowing if they're in the pool or not, so aren't sure
whether we should be pointing to pool_profile.php or participant_profile.php.
Now, the obvious *right* thing to do is combine those two files so that we're
always pointing to the same one, and different elements are shown or hidden
depending on whether a person is in the pool or not.  That change isn't really
part of this branch, though, so I'll set it aside for the moment.  The way we
had been solving the problem was to look in set_participant_id.php to see
whether a person was in the pool or not, then echo the correct path, and use
"window.location" to go to that file.  Now that we're including
core/tools/auth.php, which includes a javascript function, in
set_participant_id.php, the path echoed back is not the correct one.  So this
commit splits the path and uses just the correct piece of it (discarding the
echoed javascript) as the destination of window.location.