#78: Convert bickerdike ajax directory.
Add access checks at the top of all files in the Bickerdike ajax/ directory.
All are limited to at least data entry users, with those that allow deletions
being limited to admin users.  I will have to go back and look for any posts and
hidden elements, in order to update those with the new functions.  This is
purely access-limiting.