#78: Update logout.
Enlace and TRP each have site-specific logout procedures, but they simply
consist of closing cookies (that is, setting their values equal to zero and
setting them to expire in the past).  This commit removes those specific logout
setups (on subsite index.php pages) and replaces them with an extra line in
pleaseLogout() that unsets all existing cookies.

These cookies exist only for navigational purposes, and include for example a
"program" cookie that sets which program profile will be shown when users
navigate to it.  These cookies could conceivably all be replaced with "GET"s,
but haven't been yet.  @cwebber, thoughts on this?  I know you generally prefer
not to use cookies, which makes perfect sense.  Are GET's any better, though?