#78: Merge in changes from master (incomplete).
This commit merges in master to the Issue-78-personalized branch, resulting in
far fewer conflicts than I expected!

The remaining conflict is in /include/generalized_download_script.php, and it
comes in two parts.  One is that including the setup_user.php page also includes
the javascript failAlert() function, which then shows up on all downloads.
Obviously that will not do.  The other is that the Enlace dosage exports, one of
which was added in master after the Issue-78-personalized branch split off, are
now broken for unclear reasons.

We still need to search for $_COOKIE['user'] in all files.  It definitely still
exists in the generalized_download_script.php file, and may show up in other
places, too.

So, generally: don't expect this branch to work at the moment!  I still need to
clean up the export file.  This commit is just to complete the merge and make
note of where we stand post-merge.