#78: Remove final calls to user cookie.
Remove all references to $_COOKIE['user'] and replace them with the new API.
This required some large changes to access checks, but that's the gist of most
of this.

Using the new API caused some problems in the generalized_download_script, so I
added the ob_start() and ob_end_clean() calls.  The exports are created by
printing output, and the newly included files had some printed content (notably
a Javascript function) that was interfering in the exports.  The ob_end_clean()
call takes care of that.

I fixed the dosage calls by removing the inclusion of the new access files in
the dosage_percentage.php file.  We may need to rethink that in the future.