#119: Add and sort program-specific involvement.
Complete #119.  Added back the program-specific table to this report, with the
added bonus of filtering it by date, just as the general table does.  The date
filtering and clearing now work correctly.  Users wanted to be able to filter
the "all participants" line, but I don't think we have a date column in the
participants table, so this commit removes that line when searching by date (to
avoid confusion).