#5: Filter La Casa report by term
Add year and term filtering to the La Casa report.  It doesn't work on the race
and hometown sections yet, since those are held in the Participants table and
the query is a bit more complicated to sort those by date of entry into La
Casa.  I changed the denominators and the queries to account for the dates to
filter by.

I haven't escaped the $_POST['year_filter'] value because I was having a
surprising result (always an empty string).  I'll fix that in the next commit.

I reordered the terms in the import file as part of this commit.