#123: Remove "20-20" from schools without years.
This fixes the display of the schools that don't have years.  It adds a
conditional check of whether or not to show the years at all, depending on
whether they are available.  Furthermore, it checks whether a school has already
been accounted for via the PM_Years table, and only shows its name (without
year) if it had not already been shown.

The longer comment on this is that currently, PMs are linked to schools in two
different ways (bad!).  One way is via the Institutions_Participants table's
Is_PM column, and the other is in the PM_Years table.  The PM_Years table
generally includes a year, and the Institutions_Participants table never does.
The reason that some information wasn't showing on the participant profile is
because it had only been checking the PM_Years table, while the program profile
checked the Institutions_Participants table.