#5: Update import script with new fields
This commit adds a few new fields to the La Casa table create statements, but
mostly rewrites the import script for La Casa data.  I refactored the import
script so that each query is written in its own function, so that it'll be
easier to update the script for new years, or if the import csv's columns are
reordered.  I think there's still a ways to go on making this truly beautiful,
but we're probably only going to use it this once, so I'll stop here.

The import script now imports all non-calculated fields from the original excel
file.  Next commits will add display for these fields, and calculations to
reproduce the calculated fields in that file.